Rewrite config.m4 to use pkgconfig to locate krb5 dependencies
The config.m4 file in here used only the most rudimentary mechanisms
to detect and link to the MIT Kerberos libraries. Because of this,
it did not tolerate changes to the MIT Kerberos stack, including
when libraries like libdyn were removed.

Now the config.m4 file uses pkgconfig to get information about the
krb5 dependencies it needs. This should remain relatively future-proof,
barring any particularly drastic changes to how MIT Kerberos is organized.